requirements.txt
streamlit
streamlit-autorefresh
pandas
requests
plotly
scikit-learn
python-dotenv
